HZN-793: Reset stroke properties on each redraw
