+++ features/topology-map/org.opennms.features.topology.app/src/main/java/org/opennms/features/topology/app/internal/gwt/client/VTopologyComponent.java
@@ -277,7 +277,7 @@ public class VTopologyComponent extends Composite implements SVGTopologyMap, Top
             D3.d3().zoomTransition(selection, width, height, p0, p1);
 
             D3.d3().selectAll(GWTEdge.SVG_EDGE_ELEMENT)
-                    .style("stroke-width", GWTEdge.EDGE_WIDTH/transform.getA() + "px")
+//                    .style("stroke-width", GWTEdge.EDGE_WIDTH/transform.getA() + "px")
                     .transition()
                     .delay(750)
                     .duration(500)
